+++ e2e/tests/types.ts
@@ -6,6 +6,7 @@ export type renderTestType = {
   size: [number, number];
   percentage: number;
   title: string;
+  snapshotSuffix?: string;
   golden: string;
   only?: boolean;
   disabled?: 'node' | 'browser' | boolean;