+++ e2e/tests/visual-output/rendering/index.spec.ts
@@ -22,6 +22,11 @@ test.describe('VISUAL RENDERING TESTS', () => {
     }
 
     test(testCase.title, async ({ page }, config) => {
+      const goldenName = testCase.golden || `${testCase.title}.png`;
+      const snapshotName = testCase.snapshotSuffix
+        ? [testCase.snapshotSuffix, goldenName]
+        : goldenName;
+
       if (testCase.disabled !== 'browser') {
         await test.step(`browser - ${testCase.title}`, async () => {
           // enable and disable this inside the loop
@@ -30,13 +35,12 @@ test.describe('VISUAL RENDERING TESTS', () => {
             (testTitle) => renderingTestMap.get(testTitle)(),
             testCase.title,
           );
-
           if (output) {
             expect(
               dataURLOutputToBuffer(output),
               `browser snapshot`,
             ).toMatchSnapshot({
-              name: testCase.golden || `${testCase.title}.png`,
+              name: snapshotName,
               maxDiffPixelRatio: testCase.percentage,
             });
           } else {
@@ -44,7 +48,7 @@ test.describe('VISUAL RENDERING TESTS', () => {
               await new CanvasUtil(page).screenshot(),
               `browser snapshot`,
             ).toMatchSnapshot({
-              name: testCase.golden || `${testCase.title}.png`,
+              name: snapshotName,
               maxDiffPixelRatio: testCase.percentage,
             });
           }
@@ -66,7 +70,7 @@ test.describe('VISUAL RENDERING TESTS', () => {
               dataURLOutputToBuffer(output),
               `node snapshot should match browser snapshot`,
             ).toMatchSnapshot({
-              name: testCase.golden || `${testCase.title}.png`,
+              name: snapshotName,
               maxDiffPixelRatio: testCase.percentage,
             });
           } else {
@@ -76,7 +80,7 @@ test.describe('VISUAL RENDERING TESTS', () => {
               buffer,
               `node snapshot should match browser snapshot`,
             ).toMatchSnapshot({
-              name: testCase.golden || `${testCase.title}.png`,
+              name: snapshotName,
               maxDiffPixelRatio: testCase.percentage,
             });
           }