+++ test/visual/index.ts
@@ -1,5 +1,4 @@
 import './freedraw';
-import './group_layout';
 import './resize_filter';
 import './svg_import';
 import './text';